Move DTOs to web/rest/dto
