+++ src/main/java/de/tum/in/www1/artemis/web/rest/dto/ExamDTO.java
@@ -1,19 +1,18 @@
-package de.tum.in.www1.artemis.domain.exam.dto;
+package de.tum.in.www1.artemis.web.rest.dto;
 
+import java.io.Serializable;
 import java.time.ZonedDateTime;
 import java.util.Set;
 
 import de.tum.in.www1.artemis.domain.Course;
 import de.tum.in.www1.artemis.domain.ExerciseGroup;
-import de.tum.in.www1.artemis.domain.exam.StudentExam;
 
-public class ExamInstructorDTO {
+public class ExamDTO implements Serializable {
 
-    public ExamInstructorDTO(Long id, Course course, Set<StudentExam> studentExams, Set<ExerciseGroup> exerciseGroups, ZonedDateTime releaseDate, ZonedDateTime dueDate,
-            String startText, String endText, String confirmationStartText, String confirmationEndText, Long maxScore) {
+    public ExamDTO(Long id, Course course, Set<ExerciseGroup> exerciseGroups, ZonedDateTime releaseDate, ZonedDateTime dueDate, String startText, String endText,
+            String confirmationStartText, String confirmationEndText) {
         this.id = id;
         this.course = course;
-        this.studentExams = studentExams;
         this.exerciseGroups = exerciseGroups;
         this.releaseDate = releaseDate;
         this.dueDate = dueDate;
@@ -21,14 +20,11 @@ public class ExamInstructorDTO {
         this.endText = endText;
         this.confirmationStartText = confirmationStartText;
         this.confirmationEndText = confirmationEndText;
-        this.maxScore = maxScore;
     }
 
     public Long id;
 
     public Course course;
-
-    public Set<StudentExam> studentExams;
 
     public Set<ExerciseGroup> exerciseGroups;
 
@@ -43,6 +39,4 @@ public class ExamInstructorDTO {
     public String confirmationStartText;
 
     public String confirmationEndText;
-
-    public Long maxScore;
 }
