+++ src/main/java/de/tum/in/www1/artemis/web/rest/dto/ExamInstructorDTO.java
@@ -1,18 +1,19 @@
-package de.tum.in.www1.artemis.domain.exam.dto;
+package de.tum.in.www1.artemis.web.rest.dto;
 
-import java.io.Serializable;
 import java.time.ZonedDateTime;
 import java.util.Set;
 
 import de.tum.in.www1.artemis.domain.Course;
 import de.tum.in.www1.artemis.domain.ExerciseGroup;
+import de.tum.in.www1.artemis.domain.exam.StudentExam;
 
-public class ExamDTO implements Serializable {
+public class ExamInstructorDTO {
 
-    public ExamDTO(Long id, Course course, Set<ExerciseGroup> exerciseGroups, ZonedDateTime releaseDate, ZonedDateTime dueDate, String startText, String endText,
-            String confirmationStartText, String confirmationEndText) {
+    public ExamInstructorDTO(Long id, Course course, Set<StudentExam> studentExams, Set<ExerciseGroup> exerciseGroups, ZonedDateTime releaseDate, ZonedDateTime dueDate,
+            String startText, String endText, String confirmationStartText, String confirmationEndText, Long maxScore) {
         this.id = id;
         this.course = course;
+        this.studentExams = studentExams;
         this.exerciseGroups = exerciseGroups;
         this.releaseDate = releaseDate;
         this.dueDate = dueDate;
@@ -20,11 +21,14 @@ public class ExamDTO implements Serializable {
         this.endText = endText;
         this.confirmationStartText = confirmationStartText;
         this.confirmationEndText = confirmationEndText;
+        this.maxScore = maxScore;
     }
 
     public Long id;
 
     public Course course;
+
+    public Set<StudentExam> studentExams;
 
     public Set<ExerciseGroup> exerciseGroups;
 
@@ -39,4 +43,6 @@ public class ExamDTO implements Serializable {
     public String confirmationStartText;
 
     public String confirmationEndText;
+
+    public Long maxScore;
 }
